Fix test coverage
